Update settings for ovs_use_veth
Nevermind the patchset update on t-h-t patch, at first i tried a
different solution in patchset 4 and then i reverted it, which created
patchset 5. So patchset 5 is the same as patchset 3, no need to update
local envs. I just want to make sure that it points to the latest, which
will eventually get merged.